+++ src/pages/mainPage/CoupleContact/CoupleContact.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import ProgressHeader from '@/pages/onboarding/components/ProgressHeader';
+import ProgressHeader from '@/components/ui/progressheader';
 import Button from '@/componenets/Button/Button';
 import { Drawer, DrawerContent, DrawerHeader, DrawerFooter } from "@/components/ui/drawer";
 import Input from '@/componenets/Input/Input';
