+++ src/pages/mainPage/index.ts
@@ -1,0 +1,4 @@
+export * from "./CoupleContact/FirstMeet"
+export * from "./CoupleContact/MilitaryDay"
+export * from "./CoupleContact/CoupleContact"
+export * from "../../components/ui/progressheader"
