+++ src/pages/onboarding/index.ts
@@ -1,5 +1,5 @@
 export * from "./Onboarding"
-export * from "./CoupleContact/FirstMeet"
-export * from "./CoupleContact/MilitaryDay"
-export * from "./CoupleContact/CoupleContact"
-export * from "./components/ProgressHeader"
+export * from "../mainPage/CoupleContact/FirstMeet"
+export * from "../mainPage/CoupleContact/MilitaryDay"
+export * from "../mainPage/CoupleContact/CoupleContact"
+export * from "../../components/ui/progressheader"
